requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.2
django-filter==23.5
django-injector==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
injector==0.21.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
numpy==1.26.4
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
